no indexing

+++ misc/custom_dataset_classes.py
@@ -425,7 +425,7 @@ class H5Dataset(Dataset):
     def __getitem__(self, index):
         input = self.file['images'][index, :, :, :]
         print(np.array(self.file['labels'][index].item(0)))
-        return self.transform(input.astype('float32')), np.array(self.file['labels'][index].item(0))
+        return self.transform(input.astype('float32')), np.array(self.file['labels'][index].item(0), dtype=np.int64)
 
     def __len__(self):
         return self.n_images
